idSupplier rotation to AbstractServiceImpl, dao and service levels responsibilities fixing, id type to long everywhere fixing.

+++ src/task5/service/AbstractService.java
@@ -6,5 +6,5 @@ import java.util.NoSuchElementException;
 public interface AbstractService<T> {
     List<T> getAll();
     T getByName(String name) throws NoSuchElementException;
-    T getById(int id) throws NoSuchElementException;
+    T getById(long id) throws NoSuchElementException;
 }
